feat: UserChecklistQueryHandler
